+++ R/importPolarized.r
@@ -6,6 +6,7 @@
 #' @param changePolarity A logical vector or a list of logical vectors with length equal
 #'   to the number of markers.
 #' @param verbose Logical whether to show messages on import progress.
+#' @param simplify Logical. If `TRUE`, the function returns a matrix.
 #' @param ... Optional numeric vector of \code{compartmentSizes}.
 #' @details For details on the input data format, check the \code{file} with
 #'   \link{CheckDiemFormat}.
@@ -15,7 +16,8 @@
 #'   `TRUE` means that the genotypes encoded as `0` will be imported as `2`, and genotypes
 #'   encoded in the `file` as `2` will be imported as `0`.
 #' @return Returns a character matrix with rows containing individual genotypes and columns
-#'   containing markers.
+#'   containing markers. If `simplify = FALSE`, returns a list of matrices corresponding to
+#'   compartments in files.
 #' @seealso \link{diem} for determining appropriate marker polarity with
 #'   respect to a barrier to gene flow.
 #' @export
@@ -24,7 +26,8 @@
 #'   files <- system.file("extdata", "data7x3.txt", package = "diemr"),
 #'   changePolarity = c(FALSE, TRUE, TRUE),
 #'   ChosenInds = 1:6,
-#'   ChosenSites = "all"
+#'   ChosenSites = "all",
+#'   simplify = TRUE
 #' )
 #' dat
 #' #    m1  m2  m3
@@ -34,7 +37,7 @@
 #' # 4 "1" "2" "0"
 #' # 5 "2" "2" "1"
 #' # 6 "2" "2" "_"
-importPolarized <- function(files, changePolarity, ChosenInds, ChosenSites = "all", nCores = 1, verbose = FALSE, ...) {
+importPolarized <- function(files, changePolarity, ChosenInds, ChosenSites = "all", nCores = 1, verbose = FALSE, simplify = TRUE, ...) {
   ChosenSites <- resolveCompartments(files = files, toBeCompartmentalized = ChosenSites, ...)
   if (verbose) message("ChosenSites for compartments done ", Sys.time())
 
@@ -48,6 +51,7 @@ importPolarized <- function(files, changePolarity, ChosenInds, ChosenSites = "al
   } else if (!is.logical(changePolarity)) {
     stop("Argument 'changePolarity' must be a logical vector or a list of logical vectors.")
   }
+
   changePolarity <- resolveCompartments(files = files, toBeCompartmentalized = changePolarity, ...)
   if (verbose) message("changePolarity for compartments done ", Sys.time())
 
@@ -58,15 +62,9 @@ importPolarized <- function(files, changePolarity, ChosenInds, ChosenSites = "al
         return(NA)
       }
 
-      genotypes <- read.table(file = files[i], as.is = TRUE)
-      if ((nchar(genotypes[1, ]) - 1) < max(ChosenInds, length(ChosenInds))) {
-        stop("File ", files[[i]], " contains fewer individuals than the maximum index specified in ChosenInds.")
-      }
-      # select genotypes for chosen individuals and markers
-      genotypes <- as.data.frame(strsplit(unlist(genotypes), split = ""))[-1, ][ChosenInds, ChosenSites[[i]], drop = FALSE]
-      if (any(grepl("U", genotypes))) {
-        genotypes[genotypes == "U"] <- "_"
-      }
+      genotypes <- sImport(file = files[i], ChosenInds = ChosenInds)
+      genotypes <- genotypes[, ChosenSites[[i]], drop = FALSE]
+
       # polarise chosen markers
       genotypes <- apply(cbind(as.matrix(changePolarity[[i]][ChosenSites[[i]]]), t(genotypes)),
         MARGIN = 1,
@@ -78,10 +76,15 @@ importPolarized <- function(files, changePolarity, ChosenInds, ChosenSites = "al
 
   if (verbose) message("Importing all compartments done ", Sys.time())
 
-  allCompartments <- Filter(Negate(anyNA), allCompartments)
-  allCompartments <- do.call(cbind, allCompartments)
-  rownames(allCompartments) <- ChosenInds
-  colnames(allCompartments) <- paste0("m", markerLabels)
+  if (simplify) {
+    allCompartments <- Filter(Negate(anyNA), allCompartments)
+
+    allCompartments <- do.call(cbind, allCompartments)
+    if (ChosenInds[1] != "all") {
+      rownames(allCompartments) <- ChosenInds
+    }
+    colnames(allCompartments) <- paste0("m", markerLabels)
+  }
 
   return(allCompartments)
 }