+++ R/readIncludedSites.r
@@ -11,14 +11,13 @@
 #' @examples
 #' \dontrun{
 #' # Run this example in a working directory with write permissions
-#'   myo <- system.file("extdata", "myotis.vcf", package = "diemr")
-#'   vcf2diem(myo, "myo")
-#'   readIncludedSites("myo-001.txt")
-#'   readIncludedSites("myo-001.txt", ChosenSites = 2:7)
+#' myo <- system.file("extdata", "myotis.vcf", package = "diemr")
+#' vcf2diem(myo, "myo")
+#' readIncludedSites("myo-001.txt")
+#' readIncludedSites("myo-001.txt", ChosenSites = 2:7)
 #' }
-
 readIncludedSites <- function(includedSites, ChosenSites = "all") {
-  if(!grepl("CHROM", readLines(includedSites, n = 1)[[1]])){
+  if (!grepl("CHROM", readLines(includedSites, n = 1)[[1]])) {
     bed <- data.table::fread(includedSites, header = FALSE, sep = "\t", data.table = FALSE)[, c(1:2)]
     # convert to 1-based
     bed[, 2] <- bed[, 2] + 1