+++ NEWS.md
@@ -1,10 +1,12 @@
-# diemr 1.5.2
+# diemr 1.5.2, released 11-12-2025
 
 - Integrated BED-format support into internal functions.
 - Fixed bug in `rank2map` when a chromosome had only one site.
 - Added default value of 0.5 to `hybridIndex` when there are no data.
+- Added ploidy-aware hybrid index calculation to `hybridIndex`.
 - Modified `vcf2diem` to write all sites into one file when `chunk = 1L`, allow minimum number of required genotyped individuals per site, and subset *sampleNames.txt* to `ChosenInds` only.
 - Added check for the `changePolarity` argument in `importPolarized`.
+- Added internal function `resolveChosenInds`.
 - Updated README.